# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Cudium API Docs",
        "version": "1.0"
    },
    "paths": {
        "/api/admin/login": {
            "post": {
                "tags": [
                    "Admin"
                ],
                "summary": "Sign in the admin",
                "description": "allow admin to be able to sign in returns data object with token",
                "operationId": "authadminsignin",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Admin's Email Address",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "now@gmail.com"
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "Admin's Password",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "babatunde"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            }
        },
...
    },
    "security": [
        []
    ]
}